Refactor packages and impl IServiceHolder
Moved classes to new `service.data` and `service.manager` packages for better modularity.
Introduced `IServiceHolder` interface and refactored `ServiceHolder` to implement it, ensuring a cleaner contract for service handling.
Updated `ServiceManager` to leverage `IServiceHolder` for improved abstraction and flexibility.
Updated version to 2.0.0 to reflect breaking changes.
