+++ application.properties
@@ -1,6 +1,6 @@
 name        = Service Library
 description = A library providing manager for many similar services.
-version     = 1.1.0
+version     = 2.0.0
 
 url         = https://github.com/Akazukin-Team/Service-Library
